extract and streamline keys
