+++ app/models/InitiationGroup.java
@@ -3,6 +3,7 @@ package models;
 import com.google.i18n.phonenumbers.NumberParseException;
 import com.google.i18n.phonenumbers.PhoneNumberUtil;
 import com.google.i18n.phonenumbers.Phonenumber;
+import models.composite.InitiationKey;
 import play.data.validation.Constraints;
 
 import javax.persistence.*;
@@ -80,50 +81,4 @@ public class InitiationGroup {
         this.groupNumber = groupNumber;
     }
 
-    @Embeddable
-    public class InitiationKey {
-
-        private String guardianId;
-        private Long locationId;
-
-        public InitiationKey(String userId, Long locationId) {
-
-            this.guardianId = userId;
-            this.locationId = locationId;
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) return true;
-            if (o == null || getClass() != o.getClass()) return false;
-
-            InitiationKey that = (InitiationKey) o;
-
-            return guardianId.equals(that.guardianId) && locationId.equals(that.locationId);
-
-        }
-
-        @Override
-        public int hashCode() {
-            int result = guardianId.hashCode();
-            result = 31 * result + locationId.hashCode();
-            return result;
-        }
-
-        public String getGuardianId() {
-            return guardianId;
-        }
-
-        public void setGuardianId(String guardianId) {
-            this.guardianId = guardianId;
-        }
-
-        public Long getLocationId() {
-            return locationId;
-        }
-
-        public void setLocationId(Long locationId) {
-            this.locationId = locationId;
-        }
-    }
 }
