+++ app/models/Participation.java
@@ -1,8 +1,12 @@
 package models;
 
+import models.composite.ParticipationKey;
 import play.data.validation.Constraints;
 
-import javax.persistence.*;
+import javax.persistence.EmbeddedId;
+import javax.persistence.Entity;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 
 @Entity
 public class Participation {
@@ -93,49 +97,4 @@ public class Participation {
         NOT_ATTENDING, ATTENDING, HOSTING
     }
 
-    @Embeddable
-    public class ParticipationKey {
-
-        private Long eventId;
-        private String userId;
-
-        public ParticipationKey(final Long eventId, final String userId) {
-            this.eventId = eventId;
-            this.userId = userId;
-        }
-
-        @Override
-        public boolean equals(final Object o) {
-            if (this == o) return true;
-            if (o == null || getClass() != o.getClass()) return false;
-
-            final ParticipationKey that = (ParticipationKey) o;
-
-            return eventId.equals(that.eventId) && userId.equals(that.userId);
-
-        }
-
-        @Override
-        public int hashCode() {
-            int result = eventId.hashCode();
-            result = 31 * result + userId.hashCode();
-            return result;
-        }
-
-        public String getUserId() {
-            return userId;
-        }
-
-        public void setUserId(final String userId) {
-            this.userId = userId;
-        }
-
-        public Long getEventId() {
-            return eventId;
-        }
-
-        public void setEventId(final Long eventId) {
-            this.eventId = eventId;
-        }
-    }
 }
