+++ app/models/SuperUser.java
@@ -1,11 +1,13 @@
 package models;
 
+import models.composite.UserKey;
+
 import javax.persistence.*;
 
 @Entity
 public class SuperUser {
     @EmbeddedId
-    private SuKey key;
+    private UserKey key;
     @OneToOne(cascade = CascadeType.PERSIST)
     @JoinColumn(name = "user_id", insertable = false, updatable = false)
     private User user;
@@ -13,14 +15,14 @@ public class SuperUser {
     public SuperUser(User user) {
 
         this.user = user;
-        this.key = new SuKey(user.getId());
+        this.key = new UserKey(user.getId());
     }
 
-    public SuKey getKey() {
+    public UserKey getKey() {
         return key;
     }
 
-    public void setKey(final SuKey key) {
+    public void setKey(final UserKey key) {
         this.key = key;
     }
 
@@ -32,38 +34,4 @@ public class SuperUser {
         this.user = user;
     }
 
-    @Embeddable
-    public class SuKey {
-
-        private String userId;
-
-        public SuKey(String userId) {
-
-            this.userId = userId;
-        }
-
-        public String getUserId() {
-            return userId;
-        }
-
-        public void setUserId(final String userId) {
-            this.userId = userId;
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) return true;
-            if (o == null || getClass() != o.getClass()) return false;
-
-            SuKey suKey = (SuKey) o;
-
-            return userId.equals(suKey.userId);
-
-        }
-
-        @Override
-        public int hashCode() {
-            return userId.hashCode();
-        }
-    }
 }
